+++ backend/main.py
@@ -1,30 +1,12 @@
 import uvicorn as uvicorn
-from fastapi import FastAPI
 from fastapi_utils.tasks import repeat_every
 from loguru import logger
-from starlette.responses import FileResponse
-from starlette.staticfiles import StaticFiles
 
-from streams_explorer.api.routes import api
-
-# from streams_explorer.application import get_application
-from streams_explorer.core.config import API_PREFIX, APP_NAME, settings
+from streams_explorer.application import get_application
+from streams_explorer.core.config import settings
 from streams_explorer.default import setup_default
 
-# app = get_application()
-app = FastAPI(title=APP_NAME)
-
-app.include_router(api.router, prefix=API_PREFIX)
-
-app.mount(
-    "/_next/static", StaticFiles(directory="_next/static", html=True), name="static"
-)
-
-
-@app.get("/", include_in_schema=False)
-async def index():
-    return FileResponse("index.html")
-
+app = get_application()
 
 app.add_event_handler("startup", setup_default(app))
